chore(autoware_trajectory): move interpolated array from detail, remove unused header
