+++ common/autoware_trajectory/include/autoware/trajectory/path_point.hpp
@@ -15,7 +15,7 @@
 #ifndef AUTOWARE__TRAJECTORY__PATH_POINT_HPP_
 #define AUTOWARE__TRAJECTORY__PATH_POINT_HPP_
 
-#include "autoware/trajectory/detail/interpolated_array.hpp"
+#include "autoware/trajectory/interpolated_array.hpp"
 #include "autoware/trajectory/interpolator/result.hpp"
 #include "autoware/trajectory/pose.hpp"
